streamlit_app/pyproject.toml
[project]
name = "streamlit-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fast-plate-ocr[onnx]>=1.0.2",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "streamlit>=1.51.0",
    "ultralytics>=8.3.234",
]
